+++ benchmarks/bm_runner.py
@@ -184,16 +184,11 @@ class Branch(_SubParserGenerator):
     """Class for parsing and running the 'branch' argument."""
     name = "branch"
     description = (
-        "Performs the same operations as ``overnight``, but always on two commits "
-        "only - ``HEAD``, and ``HEAD``'s merge-base with the input "
-        "**base_branch**. If running on GitHub Actions: HEAD will be GitHub's "
-        "merge commit and merge-base will be the merge target. Performance "
-        "comparisons will be posted in a comment on the relevant pull request.\n"
-        "Designed "
-        "for testing if the active branch's changes cause performance shifts - "
-        "anticipating what would be caught by ``overnight`` once merged.\n\n"
-        "**For maximum accuracy, avoid using the machine that is running this "
-        "session. Run time could be >1 hour for the full benchmark suite.**\n"
+        "Benchmarks the two commits,``HEAD``, and ``HEAD``'s merge-base with the "
+        "input **base_branch**. If running on GitHub Actions: HEAD will be "
+        "GitHub's merge commit and merge-base will be the merge target. Performance "
+        "comparisons will be posted in the CI run which will fail if regressions "
+        "exceed the tolerance.\n"
         "Uses `asv run`."
     )
     epilog = (